requirements.txt
pip
fastapi
celery<6
redis
pymongo
uvicorn[standard]
python-decouple
requests